backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
python-multipart==0.0.21
aiofiles==25.1.0
qrcode[pil]==8.2
aiosqlite==0.22.1

# Testing
pytest==9.0.2
pytest-asyncio==1.3.0
httpx==0.28.1
